requirements.txt
aiohttp==3.9.5
pandas==2.2.2
pydantic==2.7.4
python-dotenv==1.0.1
tenacity==8.4.1
tqdm==4.66.4
